Begin working on an AVX2 parser
I've implemented the parser thus far as an example program so that I can
execute it directly.  At the moment, it operates on a 64K chunk, splits
it into four segments, and finds every word and line boundary within it.
I will benchmark it soon to evaluate how effective this paradigm can be.